+++ reporting/__init__.py
@@ -17,9 +17,11 @@ from mlframe.reporting.output import (
     parse_plot_output_dsl,
     BACKEND_FORMATS,
 )
+from mlframe.reporting.auto_dispatch import render_multi_target_panels
 
 __all__ = [
     "PlotOutputSpec",
     "parse_plot_output_dsl",
     "BACKEND_FORMATS",
+    "render_multi_target_panels",
 ]